Move web-specific functionality from main app
